feat: add login feature
add requirements: bidict package

database layer: updated password hashing method

added auth layer: add token_service to manage currently active users

service layer: add auth_service to manage login and tokens

api layer: add login and verify_usuário post requests

schemas: add user_schemas

add and update tests

+++ backend/src/tests/service/test_singup_service.py
@@ -3,11 +3,11 @@ from pydantic import BaseModel
 from src.schemas.response import HTTPResponses, HttpResponseModel
 from src.db.__init__ import user_database_example as db
 from src.db.user_database import User, UserDatabase
-from src.schemas.signup_response import HTTPSignUpResponses
+from src.schemas.user_response import HTTPSignUpResponses
 import src.service.impl.signup_service as signup
 import src.db.user_database as dbase
 
-def test_signup_user():
+def sign_up_user(use_user):
     db.remove_user_by_cpf("123.123.123-45")
     db.remove_user_by_cpf("000.000.000-00")
     
@@ -24,7 +24,7 @@ def test_signup_user():
     )
     res1 = signup.SingUpService.signup_user(dados_cadastrais, db)
     res2 = signup.SingUpService.signup_user(dados_cadastrais, db)
-    
+    assert res1 == HTTPSignUpResponses.SIGNUP_SUCCESSFUL()
     assert res2 == HTTPSignUpResponses.CPF_ALREADY_EXIST(["CPF", "USER"]) or res2 == HTTPSignUpResponses.USER_ALREADY_EXIST(["CPF", "USER"])
     
     dados_cadastrais = signup.DadosCadastrais(
@@ -42,9 +42,15 @@ def test_signup_user():
     res3 = signup.SingUpService.signup_user(dados_cadastrais, db)
     
     assert res3 == HTTPSignUpResponses.USER_ALREADY_EXIST(["USER"])
-    
-    db.remove_user_by_cpf("123.123.123-45")
-    db.remove_user_by_cpf("000.000.000-00")
+    if not use_user:
+        db.remove_user_by_cpf("123.123.123-45")
+        db.remove_user_by_cpf("000.000.000-00")
+    else:
+        return ("123.123.123-45", "000.000.000-00", "12345XyzW")
+
+
+def test_signup_user():
+    sign_up_user(False)
 
 def test_bad_request():
     dados_cadastrais = signup.DadosCadastrais(
@@ -78,3 +84,4 @@ def test_good_request():
     
     res = signup.SingUpService.signup_user(dados_cadastrais, db)
     assert res == HTTPSignUpResponses.SIGNUP_SUCCESSFUL()
+    db.remove_user_by_cpf("000.000.000-00")
